temporary changed float and double to decimal

+++ Server/BLL/DTO/AccountCurrencySummaryDto.cs
@@ -10,6 +10,6 @@ namespace BLL.DTO
     {
         public string Name { get; set; }
 
-        public double Summary { get; set; }
+        public decimal Summary { get; set; }
     }
 }
