+++ Server/MoneyManager/Model/Common/AccountTransferModel.cs
@@ -8,8 +8,8 @@ namespace MoneyManager.WEB.Model
 
         public Guid To { get; set; }
 
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
 
-        public double Fee { get; set; }
+        public decimal Fee { get; set; }
     }
 }
