+++ Server/MoneyManager/Model/Common/DepositModel.cs
@@ -11,7 +11,7 @@ namespace MoneyManager.Model.Common
 
         public DateOnly To { get; set; }
 
-        public float Percentage { get; set; }
+        public decimal Percentage { get; set; }
 
         public decimal InitialAmount { get; set; }
